chore(avm): split up AVM test contract as it was growing too large (#5702)
Attempted to add new test functions for nested contract calls and the contract grew too large (all tests began failing on deployment).

This is prep for #4293

The only "new" code in this PR is flagged as new and just adds test cases that exemplify how the kernel fails to silo side-effects by the right contract address when nested calls are made to different contracts.