feat!: migrate public to avm simulator (#6448)
This PR migrates the public (execution) environment to use the AVM
simulator. The idea of this PR is to be as minimal as possible, as to
enable easy rollbacks if needed. If things go well, there is a lot of
cleanup to do afterwards, which we are tracking in [this
issue](https://github.com/AztecProtocol/aztec-packages/issues/5818).

Major Changes
* `PublicContext` gets replaced with what was the `AvmContext`.
* Noir Aztec macros now always take the AVM path for public.
* Migration notes are added.

Other changes
* Delegate call tests are disabled, since the AVM doesn't support
delegate calls.
* ACIR public execution tests are disabled.
* Fees were changed in `bench_tx_size_fees` since the cost is now
different (bytecode and L2 gas changes).

---------

Co-authored-by: dbanks12 <david@aztecprotocol.com>