feat: store shared mutable hash (#7169)
This PR changes public write functions so that they not only store the
new delay and/or value, but also a hash of the combined serialization of
them. This lets us produce smaller proofs in private, since we only need
to prove inclusion of the hash and not each of the individual 4 values.
This will result in even larger savings once we do #5491.

The only annoying bit is that producing the unconstrained hint so that
we can contrain the hash and prove inclusion involves performing reads,
and hence computation of storage slots. But because we cannot pass
mutable references to unconstrained functions, we cannot have methods
for the `&mut PrivateContext` impl, resulting in a bit of a hack to
create a dummy state variable that I can call methods on. This is all
going to go away regardless once #5492 is done.

I also restored the tests deleted in #6985, updating the existing ones
and adding new cases.

---------

Co-authored-by: LHerskind <lasse.herskind@gmail.com>