chore: introduce execution oracle (#7521)
Discussing this and related topics with @Thunkar we realized that the
`getContractAddress` and similar oracles were duplicated: they existed
both in `UnconstrainedContext` and in `cheatcodes`. In my mind,
`cheatcodes` are oracles that only exist on TXE and that should not be
relied on for regular applications (i.e. they are not part of the PXE
API), whereas `UnconstrainedContext` _does_ use the regular PXE API.

These are clearly sort of standard, as can be seen in the fact that
we're using them outside of `UnconstrainedContext`, so I removed them
from both the context and cheatcodes, and created a new
`oracle::execution` mod with these four oracles.