refactor: add uint (U128) note to aztec-nr and remove OwnedNote from ValueNote (#8142)
In this PR we are just adding a new note type to aztec-nr. This is
simply tokennote, but we are doing the addition of this note and the
removal of tokennote in two steps to make the diffs easier to
understand. This also removes `OwnedNote` from `ValueNote`.

Sets up resolving of #8107.