feat: hash contract class logs with poseidon + add preimage to blob (#12061)
~Quick and dirty~ Somewhat clean implementation which:

- Converts our SHA based contract class `LogHash`es with poseidon hashes
and constrain them in the registerer.
- Verifies this ^ hash in the base rollup and appends to the blob as
fields, eliminating the need for any calldata.
- Treats the logs as fields in ts instead of bytes.

EDIT: Added log validation to tx in this commit:
https://github.com/AztecProtocol/aztec-packages/pull/12061/commits/cb739a21aca3719ba75b41f7a68cd268bd09f448

EDIT 2: We may not need to constrain the hash in the registerer, so can
revert to the oracle hashing the log and leaving the constraining to the
base rollup in a future PR.

---------

Co-authored-by: Leila Wang <leizciw@gmail.com>