chore: rename logs to msg (#13364)
This expands on the 'message' term:

- contracts send and receive messages
- messages follow standard encoding, by having a message type id,
metadata, and content
- encoded messages are encrypted into ciphertext
- ciphertext can be emitted alongside a tag in a log
- ciphertext is decrypted into plaintext, which is an encoded message
- from it we get back message metadata and content

Before this PR, we call both the log and the message 'logs', which is
incorrect and confusing.

The renaming is partial since some concepts are still tied to logs (e.g.
events *must* be in logs, message processing is currently done on logs
stored in capsules, etc.), but this is a good first step.