feat!: use given lengths to trim emitted logs (#14041)
- All types of logs (private, public, contract class) will now be
emitted from the contracts with a length. The base rollup will include
the logs to the blobs based on the specified length.

Refactoring:
- Change the prefixes of side effects in blobs to be the number of items
instead of the total fields of all items. It's cheaper to compute and
easier to deserialise.
- Remove `counter` from `LogHash` when outputted from private tail.
Counter is only required when processing in private. We used to set it
to 0 when exposing to public.

Renaming:
- Rename `[PUBLIC/CONTRACT_CLASS]_LOG_DATA_SIZE_IN_FIELDS ` to
`[PUBLIC/CONTRACT_CLASS]_LOG_SIZE_IN_FIELDS ` to be consistent with
`PRIVATE_LOG_SIZE_IN_FIELDS`.
- Rename `[PUBLIC/CONTRACT_CLASS]_LOG_SIZE_IN_FIELDS ` to
`[PUBLIC/CONTRACT_CLASS]_LOG_LENGTH` because other constants for a
struct's total number of fields are named `[...]_LENGTH`.